backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
twilio==9.4.1
pypdf==5.1.0
python-docx==1.1.2
